Code improvements and warning for unsupported resources (#65)

+++ internal/broker/utilities.go
@@ -209,7 +209,7 @@ func client(providerData *providerData) (*semp.Client, diag.Diagnostic) {
 		true, // this is a client for the provider
 		semp.BasicAuth(username, password),
 		semp.BearerToken(bearerToken),
-		semp.Retries(uint(retries), retryMinInterval, retryMaxInterval),
+		semp.Retries(retries, retryMinInterval, retryMaxInterval),
 		semp.RequestLimits(requestTimeoutDuration, requestMinInterval))
 	return client, nil
 }
